fix(analysis): address Copilot review (5 actionable findings) (#497)
PR #538 review pass — all 5 Copilot inline comments resolved:

1. NewAnalysisModal date input no longer uses ``toISOString().slice(0,10)``
   which shifted the calendar day for non-UTC users (JST users at 23:30
   would default the modal to "tomorrow"). Now uses ``formatLocalDate``
   from ``lib/utils/datetime`` — same helper the cost dashboard already
   uses for ``<input type="date">`` defaults.

2. ``AnalysesTabPanel`` KPI quality value moved from a hardcoded
   ``conf {x}`` template literal to ``t("kpi.qualityValue", {value})``.
   Adds the new key to both en.json and ja.json.

3. Extracted the pure ``check_workspace_in_allowlist`` helper into
   ``auth/analysis_allowlist.py`` so callers that only need the
   membership check (e.g. ``api/routes/workspaces`` exposing
   ``analyses_enabled``) do not pull in ``auth.analysis_gates``'s quota
   / tier / BYOK dependency tree. ``analysis_gates`` re-exports for
   backward compat — internal gate callers and tests are unchanged.

4. ``RepresentativesPanel`` meta line ("{id} · {type} · importance {n}")
   moved from a hardcoded English template into ``t("metaLine", ...)``
   with named parameters. Now translates correctly for ja users.

5. Updated ``analyses.ts`` ``AnalysisFilters`` doc comment to reference
   the actually-existing ``formatLocalDate`` helper instead of the
   nonexistent ``toUtcIso``. The wire-format guidance now matches what
   ``NewAnalysisModal`` does after fix #1.

## Quality
- Backend ruff: All checks passed
- Backend pytest: 22/22 (analyses routes + quota gate scopes)
- Frontend TypeScript: 0 errors
- Frontend Vitest: 20/20 (analyses scope, no test changes)
